backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "wu",email = "wu@123.com"}
]

packages = [
    { include = "backend" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "pyjwt (>=2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
